requirements.txt
backgroundremover==0.2.5
certifi==2023.7.22
charset-normalizer==3.3.2
commandlines==0.4.1
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.13.1
filetype==1.2.0
fsspec==2023.10.0
future==0.18.3
hsh==1.1.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
PyMatting==1.1.11
PySocks==1.7.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
waitress==2.1.2
